- added WIN64 support - added abstraction for Windows header includes

+++ HomerBase/include/HBReflection.h
@@ -36,7 +36,7 @@ inline std::string ParseRawObjectName(std::string pRawName)
 {
 	std::string tResult = "";
 	int tPos = 2; //ignore prefix "PN"
-	int tRawNameLength = pRawName.length();
+	int tRawNameLength = (int)pRawName.length();
 
 	for(;;)
 	{
